fix: refactored code to remove circular dependency
