+++ packages/shared/package.json
@@ -37,11 +37,13 @@
         "moment": "^2.29.2"
     },
     "dependencies": {
+        "@deriv/translations": "^1.0.0",
         "@types/js-cookie": "^3.0.1",
         "@types/moment": "^2.13.0",
         "@types/react-loadable": "^5.5.6",
         "canvas-toBlob": "^1.0.0",
         "extend": "^3.0.2",
+        "i18next": "^20.3.2",
         "js-cookie": "^2.2.1",
         "moment": "^2.29.2",
         "object.fromentries": "^2.0.0",
