+++ packages/shared/src/utils/config/platform-config.js
@@ -1,3 +1,7 @@
+import i18n from 'i18next';
+import React from 'react';
+import { getInitialLanguage } from '@deriv/translations';
+import { initMoment } from '../date';
 import { routes } from '../routes';
 
 // TODO: This should be moved to PlatformContext
@@ -16,4 +20,28 @@ export const platforms = {
         route_to_path: '',
         url: '',
     },
+};
+
+export const useOnLoadTranslation = () => {
+    const [is_loaded, setLoaded] = React.useState(false);
+
+    React.useEffect(() => {
+        if (!i18n.language) {
+            i18n.language = getInitialLanguage();
+        }
+        const is_english = i18n.language === 'EN';
+
+        if (is_english) {
+            setLoaded(true);
+        } else {
+            i18n.store.on('added', () => {
+                setLoaded(true);
+            });
+        }
+
+        return () => i18n.store.off('added');
+    }, []);
+
+    initMoment(i18n.language);
+    return [is_loaded, setLoaded];
 };
