+++ packages/translations/package.json
@@ -31,7 +31,6 @@
     "url": "https://github.com/binary-com/deriv-app/issues"
   },
   "dependencies": {
-    "@deriv/shared": "^1.0.0",
     "commander": "^3.0.2",
     "crc-32": "^1.2.0",
     "glob": "^7.1.5",
